introduce an Installer class to allow associating metadata with each installer
This will allow individual installer functions to attach contextual metadata
to it. This will enable more in-depth information and analysis of what will be
executed which can be leveraged in different ways (e.g., display the series of
actions that will be executed when installing a server, prior to installing
it).

It can also allow installers to control how they're executed, for example
whether a tmpdir should be used or not. This is the main motivation of
introducing it.

Fixes #223.